inject cast to avoid warning about size_t being printed as an int.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@23048 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-cpp/x10/lang/Thread.cc
@@ -210,9 +210,9 @@ Thread::thread_init(const ref<String> name)
     if (defined) {
         int rc = pthread_attr_setstacksize(&__xthread_attr, stacksize);
         if (rc != 0) {
-            ::fprintf(stderr, "Cannot set stack size to %d; %s. Using default size instead.\n", stacksize, ::strerror(rc));
+            ::fprintf(stderr, "Cannot set stack size to %d; %s. Using default size instead.\n", (int)stacksize, ::strerror(rc));
         } else {
-            ::fprintf(stderr, "Successfully set stack size to %d\n", stacksize);
+            ::fprintf(stderr, "Successfully set stack size to %d\n", (int)stacksize);
         }
     }
